chore(examples): reorder with range-based numbering convention
Category-based numbering scheme (0xx core, 1xx customization, 2xx ecosystem) so new examples slot in without renumbering.
