--- examples/08_tree-sitter-syntax-highlighting/go.mod
@@ -1,29 +1,0 @@
-module github.com/indaco/herald/examples/08_tree-sitter-syntax-highlighting
-
-go 1.25.0
-
-require (
-	charm.land/lipgloss/v2 v2.0.2
-	github.com/indaco/herald v0.0.0
-	github.com/smacker/go-tree-sitter v0.0.0-20240827094217-dd81d9e9be82
-)
-
-require (
-	github.com/charmbracelet/colorprofile v0.4.2 // indirect
-	github.com/charmbracelet/ultraviolet v0.0.0-20251205161215-1948445e3318 // indirect
-	github.com/charmbracelet/x/ansi v0.11.6 // indirect
-	github.com/charmbracelet/x/term v0.2.2 // indirect
-	github.com/charmbracelet/x/termios v0.1.1 // indirect
-	github.com/charmbracelet/x/windows v0.2.2 // indirect
-	github.com/clipperhouse/displaywidth v0.11.0 // indirect
-	github.com/clipperhouse/uax29/v2 v2.7.0 // indirect
-	github.com/lucasb-eyer/go-colorful v1.3.0 // indirect
-	github.com/mattn/go-runewidth v0.0.19 // indirect
-	github.com/muesli/cancelreader v0.2.2 // indirect
-	github.com/rivo/uniseg v0.4.7 // indirect
-	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
-	golang.org/x/sync v0.18.0 // indirect
-	golang.org/x/sys v0.42.0 // indirect
-)
-
-replace github.com/indaco/herald => ../..
